[code-infra] Add documentation to internal types
