+++ packages/x-internals/src/types/index.ts
@@ -1,17 +1,4 @@
-export type MakeOptional<T, K extends keyof T> = Omit<T, K> & Partial<Pick<T, K>>;
-
-export type MakeRequired<T, K extends keyof T> = Omit<T, K> & Required<Pick<T, K>>;
-
-export type DefaultizedProps<
-  P extends {},
-  RequiredProps extends keyof P,
-  AdditionalProps extends {} = {},
-> = Omit<P, RequiredProps | keyof AdditionalProps> &
-  Required<Pick<P, RequiredProps>> &
-  AdditionalProps;
-
-export type SlotComponentPropsFromProps<
-  TProps extends {},
-  TOverrides extends {},
-  TOwnerState extends {},
-> = (Partial<TProps> & TOverrides) | ((ownerState: TOwnerState) => Partial<TProps> & TOverrides);
+export * from './DefaultizedProps';
+export * from './MakeOptional';
+export * from './MakeRequired';
+export * from './SlotComponentPropsFromProps';
